src/pyproject.toml
[tool.poetry]
name = "webapp"
version = "0.1.0"
description = ""
authors = ["kloeffler95 <kevin.loeffler@colorado.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
requests = "^2.31.0"
squareup = "^27.0.0.20230517"
flask-qrcode = "^3.1.0"
tplink-cloud-api = "^4.1.5"
configparser = "^5.3.0"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
aiofiles = "^23.1.0"
black = "^23.3.0"
geopy = "^2.3.0"
apscheduler = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
